Initial commit
Commands:
- /ladder collect
    - Allows you to collect a ladder from any vehicle listed in the config
- /ladder delete
    - Deletes the ladder in your hand
    - This is temporary until a proper system to collect and store ladders is made.

Every other feature of the resource can be found through the in-game prompts. Tested on localhost with no major issues. Tested on the test server with one other person with no major issues. If you test this resource in your server, please report back to us! We have not thoroughly tested this update and would appreciate your feedback.

Known bugs: Carrying the ladder looks unlegit, climbing the ladder looks a bit unlegit as well, all will be improved in-time.