Add per vehicle ladders, changed networking, changed events
The delete command how now been removed, and replaced with `/ladder store`. This new command allows you to place ladders back into any vehicle that can carry ladders, and has room to do so; how many ladders a vehicle model can carry can be defined in the config. A ladder does not have to be placed back into the vehicle it was removed from, as long as it has enough room for another ladder.

Removed some networking natives that were not needed apparently; resource seems to run fine without them, soooo, cool?

Changed the client and server event names and structure, just because a lot of the stuff was the same anyway.