New ladder, one-lick climb, back-end changes
Exactly the same as the last two commits, but maybe GitHub will work this time.

DO NOT BLINDLY UPDATE // THIS IS A DEVELOPMENT BUILD

Important Note: If you test this resource in your server, please report back to us! We have not thoroughly tested this update and would appreciate your feedback.

Added:
- New, taller, metal ladder, courtesy of Adam Fenton#4173.
- One-click climb button.
    - Now when you walk up to a ladder, it determines if you are at the top or button of the ladder, and automatically lets you climb in the correct direction.

Changed:
- Climbing function to use a loop instead of a long chain.
- Master loop to increase efficiency.
- When you reach the top of a ladder it now 'throws' you a little bit further.
- When you reach the bottom of a ladder it now places you a bit better on the ground.
- Updated manifest file to latest spec.
- Most "s to 's to match new project standards.